+++ src/main/java/net/lopymine/patpat/compat/replaymod/ReplayModCompat.java
@@ -8,17 +8,16 @@ import net.lopymine.patpat.client.packet.PatPatClientPacketManager;
 import net.lopymine.patpat.compat.LoadedMods;
 import net.lopymine.patpat.packet.s2c.*;
 
-
 public class ReplayModCompat {
 
 	public static void onPat(int pattedEntityId, int whoPattedId) {
 		//? if replaymod {
-		/*if (!LoadedMods.REPLAY_MOD_LOADED) {
+		if (!LoadedMods.REPLAY_MOD_LOADED) {
 			return;
 		}
 		SelfPatEntityS2CPacketV2 packet = new SelfPatEntityS2CPacketV2(pattedEntityId, whoPattedId);
-		ReplayModManager.sendDummyPacket(ServerPlayNetworking.createS2CPacket(packet));
+		ReplayModManager.sendDummyPacket(ServerPlayNetworking.createClientboundPacket(packet));
 		PatPatClientPacketManager.LOGGER.debug("Saved packet to ReplayMod");
-		*///?}
+		//?}
 	}
 }