+++ src/main/java/net/lopymine/patpat/mixin/controlling/IPlatformHelperMixin.java
@@ -3,6 +3,7 @@ package net.lopymine.patpat.mixin.controlling;
 //? if controlling {
 
 import com.blamejared.controlling.client.NewKeyBindsScreen;
+import com.mojang.blaze3d.platform.InputConstants;
 import com.blamejared.controlling.platform.IPlatformHelper;
 import net.lopymine.patpat.client.keybinding.PatPatClientKeybindingManager;
 import net.minecraft.client.Options;
@@ -15,9 +16,7 @@ public interface IPlatformHelperMixin {
 
 	@Inject(at = @At("HEAD"), method = "handleKeyPress", cancellable = true)
 	private void handlePatPatKeybindingOnKeyPressed(NewKeyBindsScreen screen, Options options, net.minecraft.client.input.KeyEvent event, CallbackInfo ci) {
-		int keyCode = event.key();
-		int scanCode = event.scancode();
-		PatPatClientKeybindingManager.handlePatPatKeybindingOnKeyPressed(screen.selectedKey, screen, keyCode, scanCode, ci::cancel);
+		PatPatClientKeybindingManager.handlePatPatKeybindingOnKeyPressed(screen.selectedKey, screen, InputConstants.getKey(event), ci::cancel);
 	}
 
 }